Move packages under rpc
